muevo obtencion de query con filtros para getProperties a utils
